Added some minor things (in Description)
Added some comments to the Javascript file.
Added a success message for the server side that the javascript will
read to determine if it needs to go to the confirmation page.
Added a small styling that will change the color of the form if it is
valid or invalid.
other minor changes