Server Validation, MainStyling, AdminReport Page
Created a basic server validation using preg_match and the regex from
the javascript file.

Created a css for the pages, index only for now.

Created an Admin report page that will display the database info in
descending order based on when the info was added(Date Entered).

Whats left:
Make the Confirmation page look nice.